webapp/form: merge
